refactor(eval): move loc baseline runner into agent package
